requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
CTkMessagebox==2.7
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.8
Django==5.1.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.0.1
fastcli==0.2.0
fastdotcom==0.0.3
filelock==3.16.0
Flask==3.0.3
flaskwebgui==1.1.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
gast==0.6.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==2.15.0
keyboard==0.13.5
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml-dtypes==0.2.0
MouseInfo==0.1.3
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.5
nest-asyncio==1.6.0
notify2==0.3.1
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pbr==6.1.0
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
plyer==2.1.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyspeedtest==1.2.7
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
pywebio==1.8.3
pywin32==306
pywin32-ctypes==0.2.3
pyzmq==26.1.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
speedtest-cli==2.1.3
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.38.2
stevedore==5.3.0
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.2
user-agents==2.2.0
validators==0.34.0
vgamepad==0.1.0
virtualenv==20.26.4
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.14.1
yarl==1.9.6
